requirements.txt
playwright
tqdm
playwright-stealth
mysql-connector-python
python-dotenv
deutschland
setuptools
rapidfuzz
fastapi
questionary
uvicorn
ollama